model update

+++ base/models.py
@@ -5,8 +5,8 @@ from django.contrib.auth.models import User
 
 class Geolocation(models.Model):
   # user = models.ForeignKey(User, on_delete=models.SET_NULL, null=True)
-  lon = models.DecimalField(max_digits=8, decimal_places=3, null=True,)
-  lat = models.DecimalField(max_digits=8, decimal_places=3, null=True,)
+  lon = models.FloatField(null=True)
+  lat = models.FloatField(null=True)
   updated = models.DateTimeField(auto_now=True)
   created = models.DateTimeField(auto_now_add=True)
